Clarify explanation of scalar/vector fields on page 5
The paragraph stated that a scalar field outputs a single numeric value... but didn't
say what it takes as *input*. Any garden-variety function like x^2 outputs a numeric
value, but that doesn't make it a scalar field.

Specify that it "takes a point in space as input" to make the definition complete.